front page: change 6 panel img to background css
Added CSS to support changing the markup in the home page from using img elements for the six images used in topic panels on the Wordpress home page to be css background images instead.  This allows narrow-width browsers e.g. mobiles (a layout which does not display the images anyway) to avoid downloading the image files.